Add tests
We add a significant open-source project (APR) to be built with gcc-wrapper.
Once *._i_.c files are obtained we replace original files with produced ones
and try to compile the whole project from scratch.
The test is considered successful if we've managed to built the project
from produced files taken in place of original C source files.